backend/requirements.txt
Django==4.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
django-filter==23.5
psycopg2-binary==2.9.9
python-dotenv==1.0.0
Pillow==10.1.0

# Celery and Redis
celery==5.3.4
redis==5.0.1

# ML/AI Libraries
scikit-learn==1.3.2
pandas==2.1.4
numpy==1.26.2
spacy==3.7.2
prophet==1.1.5

# Testing
pytest==7.4.3
pytest-django==4.7.0

# Production server
gunicorn==21.2.0

# PDF Processing
pdfplumber==0.10.3
PyPDF2==3.0.1
